requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
colorama==0.4.3
curl==0.0.1
Django==3.0.6
django-rest-auth==0.9.5
djangorestframework==3.11.0
httpie==2.1.0
idna==2.9
Pygments==2.6.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
